% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnorm.R
\name{cnorm}
\alias{cnorm}
\title{Check Normality of Data}
\usage{
cnorm(x)
}
\arguments{
\item{x}{A numeric vector for which normality is to be checked.}
}
\value{
A list containing the following elements:
  \itemize{
    \item \code{histogram}: A histogram of the data.
    \item \code{qq_plot}: A Q-Q plot of the data.
    \item \code{shapiro_test}: Results of the Shapiro-Wilk test.
  }
}
\description{
This function checks the normality of a given numeric vector using visual
inspection (histogram and Q-Q plot) and a statistical test (Shapiro-Wilk).
}
\examples{
data <- rnorm(100)
cnorm(data)

}
